requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.10
celery==5.3.4
redis==5.0.1
python-decouple==3.8
langchain>=0.3.0
langchain-core>=0.3.0
langchain-huggingface==0.3.1
huggingface-hub==0.33.4
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
gunicorn==21.2.0
dj-database-url==2.1.0
tiktoken>=0.8.0
